generify MixtureDataset a bit
